pyproject.toml
[tool.poetry]
name = "homework-6"
version = "0.1.0"
description = ""
authors = ["Denys Kotsiuba"]
readme = "README.md"
packages = [{include = "homework_6"}]

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.6"
faker = "^18.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
